Almost clened up code
