+++ src/assets/.d.ts
@@ -1,4 +1,4 @@
-declare module "*.svg" {
-  const content: any
-  export default content
-}
+// declare module "*.svg" {
+//   const content: any
+//   export default content
+// }
