+++ src/components/ErrorBoundary.tsx
@@ -1,20 +1,19 @@
 import { ErrorBoundary } from 'react-error-boundary'
 
 const ErrorFallback: React.FunctionComponent<{
-    readonly error: Error
-}> = ({ error }) => {
-    return (
-        <div role="alert">
-            <p>Something went wrong:</p>
-            <pre>{error.message}</pre>
-        </div>
-    )
-}
+  readonly error: Error
+}> = ({ error }) =>
+  (
+    <div role="alert">
+      <p>Something went wrong:</p>
+      <pre>{error.message}</pre>
+    </div>
+  )
 
 const Main: React.FunctionComponent<React.PropsWithChildren> = ({
-    children,
+  children,
 }) => (
-    <ErrorBoundary FallbackComponent={ErrorFallback}>{children}</ErrorBoundary>
+  <ErrorBoundary FallbackComponent={ErrorFallback}>{children}</ErrorBoundary>
 )
 
 export default Main
