+++ src/components/ErrorMessage.tsx
@@ -1,6 +1,5 @@
 export const ErrorMessage: React.FunctionComponent<{
   errorMessage?: string
-
 }> = ({ errorMessage }) =>
   (
     <div className="error-message">{errorMessage}</div>
