+++ src/components/Footer.tsx
@@ -4,7 +4,7 @@ export const Footer: React.FunctionComponent = () =>
         <footer className="footer">
             <ul className="footer-ul">
                 <li className="footer-li">&copy; Giang Dinh 2023</li>
-                <li className="footer-li">Giadi735@gmail.com</li>
+                <li className="footer-li">giadi735@gmail.com</li>
                 <li className="footer-li">
                     <a href="https://www.linkedin.com/in/giang-dinh-99686ba6/" className="footer-link">
                         LinkedIn
@@ -12,7 +12,7 @@ export const Footer: React.FunctionComponent = () =>
                 </li>
                 <li className="footer-li">
                     <a href="https://github.com/fruddq" className="footer-link">
-                        Github
+                        github
                     </a>{" "}
                 </li>
             </ul>
