+++ src/components/Menu.tsx
@@ -1,6 +1,5 @@
 import { useCallback, useState } from "react"
 import { useNavigate } from "react-router-dom"
-// import { navigate } from "wouter/use-location"
 import { Dropdown } from "./Dropdown"
 import { SearchSVG } from "./SearchSVG"
