+++ src/components/Movie.tsx
@@ -1,7 +1,6 @@
 import CN from 'classnames'
 import { useCallback, useContext, useEffect, useState } from "react"
 
-import type { PropsMovie } from "../models/Props"
 import { genreList } from "../services/config"
 import { EActionFavorites, storeFavorites } from "../stores/favorites"
 import { EActionWatchLater, storeWatchLater } from '../stores/watchLater'
@@ -9,10 +8,15 @@ import { EActionWatchLater, storeWatchLater } from '../stores/watchLater'
 import placeHolder from '../assets/place-holder.png'
 import { Trailers } from './Trailers'
 import { fetchTrailers } from '../services/fetchTrailers'
+import type { IMovie } from '../models/Interfaces'
 
 const getTrailers = async (movieID: number) => {
   const trailerKeys = await fetchTrailers(movieID)
   return (trailerKeys.map(trailer => `https://www.youtube.com/embed/${trailer}`))
+}
+
+export interface PropsMovie {
+  movie: IMovie
 }
 
 export const Movie: React.FunctionComponent<PropsMovie> = ({ movie }) => {
